requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras
#
annotated-types==0.6.0
    # via pydantic
click==8.1.7
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via pytest
pydantic==2.5.2
    # via -r requirements.in
pydantic-core==2.14.5
    # via pydantic
pytest==7.4.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tabulate==0.9.0
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
